server/requirements.txt
django-cors-headers
django-filter
django
djangorestframework
gemmi
pytest
requests-cache
requests
uvicorn
whitenoise
